+++ lib/src/extensions/device_extensions.dart
@@ -32,7 +32,10 @@ bool get isGoogle => !isWeb && (Platform.isAndroid || Platform.isFuchsia);
 
 bool get isAndroid => !isWeb && Platform.isAndroid;
 
+@Deprecated('Use isIOS instead')
 bool get isIos => !isWeb && Platform.isIOS;
+
+bool get isIOS => !isWeb && Platform.isIOS;
 
 bool get isMacOS => !isWeb && Platform.isMacOS;