+++ lib/src/utils/text_styles.dart
@@ -103,6 +103,7 @@ TextStyle secondaryTextStyle({
 }
 
 // Create Rich Text
+@Deprecated('Use RichTextWidget instead')
 RichText createRichText({
   required List<TextSpan> list,
   TextOverflow overflow = TextOverflow.clip,